[hw1] Novozhilov Dmitriy. Move `printer_translator.h` in `includes\`
